revise login/home page in frontend
